+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## Version 0.0.2
+MagicPanel -> Fixed RPC Error if Text is null
+MessagePanel -> Don't send null text if only 1 message
+
 ## Version 0.0.1
 Inital Test Version